tower upgrade pop up window adjusted >> can show the cost and attack of each stage of the tower bullet speed adjusted to "hit matches hp decline" money or hp changes, color of the font changes

+++ src/controllers/TowerController.java
@@ -21,8 +21,8 @@ public class TowerController {
     private LinkedList<Tower> towers;
     private LinkedList<Alien> aliens;
     private LinkedList<Point> range;
-    public static final int costArr[] = {TOWER0_COST, TOWER1_COST, TOWER2_COST, TOWER3_COST, TOWER4_COST};
-    public static final int upgradeCostArr[] = {TOWER0_COST/2, TOWER1_COST/2, TOWER2_COST/2, TOWER3_COST/2, TOWER4_COST/2};
+    public static final int costArr[] = {(int)TOWER0_COST, (int)TOWER1_COST, (int)TOWER2_COST, (int)TOWER3_COST, (int)TOWER4_COST};
+    public static final int upgradeCostArr[] = {(int)TOWER0_COST/2, (int)TOWER1_COST/2, (int)TOWER2_COST/2, (int)TOWER3_COST/2, (int)TOWER4_COST/2};
 
     public TowerController(AlienController alienController) {
         towers = new LinkedList<Tower>();
