+++ src/values/Global.java
@@ -43,16 +43,16 @@ public class Global {
     public static final float TOWER2_ATKRANGE = 3 * SIZE_GRID;
     public static final float TOWER3_ATKRANGE = 4 * SIZE_GRID;
     public static final float TOWER4_ATKRANGE = 5 * SIZE_GRID;
-    public static final int TOWER0_COST = 40;
-    public static final int TOWER1_COST = 60;
-    public static final int TOWER2_COST = 80;
-    public static final int TOWER3_COST = 100;
-    public static final int TOWER4_COST = 120;
-    public static final int TOWER0_ATK = 20;
-    public static final int TOWER1_ATK = 30;
-    public static final int TOWER2_ATK = 50;
-    public static final int TOWER3_ATK = 60;
-    public static final int TOWER4_ATK = 70;
+    public static final float TOWER0_COST = 40;
+    public static final float TOWER1_COST = 60;
+    public static final float TOWER2_COST = 80;
+    public static final float TOWER3_COST = 100;
+    public static final float TOWER4_COST = 120;
+    public static final float TOWER0_ATK = 20;
+    public static final float TOWER1_ATK = 30;
+    public static final float TOWER2_ATK = 50;
+    public static final float TOWER3_ATK = 60;
+    public static final float TOWER4_ATK = 70;
 
     //Alien
     public static float [] START_POINT1 = {-SIZE_GRID, SIZE_GRID*2};
